requirements.txt
streamlit
time
datetime
fitter
pandas
numpy
matplotlib.pyplot
pickle
yfinance
streamlit-aggrid